Hide problem weight when set to 0
This commit hides the problem weight when the problem
weight is set to zero. This makes the free-text-xblock
follow the convention used by problem types, which also
hide the "0 points possible" text instead of displaying
it. A test of this feature has been added to the suite
of tests for this xblock.