Add cache to display answer after submit or save
Previously in the free-text response xblock, when a user submits or
saves an answer, navigates to another page, and navigates back to the
original page, the user's answer disappears. Refreshing the webpage will
restore the answer, but this was nonetheless a disconcerting experience
for the user.

This fix caches the user's answer to display after submit or save, so
that when the user navigates away and back, the user can still view
their answer without having to manually refresh the page.